+++ packages/doctor/src/index.ts
@@ -9,13 +9,13 @@ import { IFileInfo } from './types/File';
 // Ignore directories
 const defaultignore = ['.faas_debug_tmp', '.ice', '.rax', 'build', 'es', 'dist', 'lib', 'mocks', 'logs', 'coverage', 'node_modules', 'demo', 'examples', 'public', 'test', '__tests__'];
 class Doctor {
-  options: any;
+  options: IDoctorOptions;
 
   ignore: string[];
 
-  private scanner: any;
+  private scanner: Scanner;
 
-  private analyzer: any;
+  private analyzer: Analyzer;
 
   constructor(options: IDoctorOptions) {
     this.options = options || {};